requirements.txt
matplotlib==3.9.0
moviepy==1.0.3
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.2
pillow==10.3.0
psutil==5.9.8
requests==2.32.2
scipy==1.13.0
ultralytics==8.2.19
# git+https://github.com/Noah-Grimaldi/faster-whisper.git # if you are on MAC, make sure you are using Python3.10

setuptools~=70.0.0
rich~=13.7.1
